HashiCorp Vault integration and updated dependencies (#23)
* pkg/httpclient: initiate insecure connection only when explicitly asked

* Dockerfile: Update Golang to v1.20

* build: generate PIE binaries

* Makefile: Update build tools

Update kustomize, protoc, controller-gen to the latest releases.

Also download and use protobuf plugins for generating the sources.

* build: update dependencies

* Dockerfile: add vendor to wrokspace

Use the vendor tree generated on the host.

* isecl/kmiptool: load configuration from file

* kmip: configure option for setting log level

* plugins/isecl: changes to support kmip v1.4 spec

* plugins/isecl: fetch singer certificate only if KMIP is configured

In case the isecl plugins is configured with no KMIP server, do not
treat it as error. Instead, ignore the signer certificate fetching. This
will be useful for the key-only fetching setup like Vault.

* isecl/kbs-tool: improved command usage

* plugin/isecl/kbs-tool: read KBS configuration from file

* docs: How to document for using Vault as key server